📝 Modularize authentication middleware
