+++ forecast/server/src/middleware/index.js
@@ -1,0 +1,4 @@
+const authenticate = require('./authenticate')
+const sesh = require('./session')
+
+module.exports = { authenticate, sesh }
